[1.4] For libogg use Windows 10 SDK version 10.0.17763.0
What is the correct way to replace this?

There is WindowsSDKVersion as an environment variable. But that has a
trailing backslash. And where does this even come from?

There is also env var UCRTVersion, but it is unclear what that declares exactly.

We may want to target the Windows 8.1 SDK anyway. That would mean, with
this patch not applied, the user will have to install the Windows 8.1 SDK.